read before use
A project based on android which provide the online booking of room for pg/hotel /apartment . The Authentic user can add room to this app which is shown to the user . The Jetflix of mobile technology opens the windows to the android app. The websites are vanishing and the mobile phones are emerging. It’s the time to change from conventional websites to apps, which has become the part of our daily routine. we are introducing Jetflix.apk‟ the android application software which would be solving the problem of room booking in the University .Not only we can use this app in university but we can use this app anywhere. It works as the guidance for students by providing the details of free available rooms nearby .After searching for room one can find their detail and also contact the owner of room . This project gives a total solution to everyone student like a hub. It gives us more comfort and a better user interface . The student can easily book room as well as contact the room provider through the app .The google map will show the location of room thus provide the way to reach the room easily. It acts as an overview about the digital campus. Individual account can be created for user just to use the features of rating and reviewing. The trusted member can add room for PG , hotel , Apartment etc which will be shown to the user at runtime .